+++ src/main/java/com/example/challengebackendjava/controller/PersonajeController.java
@@ -7,12 +7,12 @@ import com.example.challengebackendjava.service.PeliculaSerieService;
 import com.example.challengebackendjava.service.PersonajeService;
 import com.fasterxml.jackson.annotation.JsonView;
 import io.swagger.annotations.ApiOperation;
-import io.swagger.models.auth.In;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.*;
+import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @RestController
@@ -29,7 +29,7 @@ public class PersonajeController {
   public ResponseEntity<List<Personaje>> getPersonajes(
       @RequestParam(required = false) String name,
       @RequestParam(required = false) Integer age,
-      @RequestParam(required = false) Set<Integer> idMovies
+      @RequestParam(required = false) Set<Long> idMovies
       ) {
 
     List<Personaje> personajes = personajeService.all();
@@ -58,13 +58,13 @@ public class PersonajeController {
   @GetMapping("/characters/{id}")
   @JsonView(View.Personaje.Detalle.class)
   @ApiOperation("Permite buscar un personaje por su id, con toda su informacion")
-  public ResponseEntity<Personaje> getPersonaje(@PathVariable Integer id) {
+  public ResponseEntity<Personaje> getPersonaje(@PathVariable Long id) {
     return ResponseEntity.ok(personajeService.findById(id));
   }
 
   @PutMapping("/characters/{id}")
   @ApiOperation("Permite actualizar un personaje por su id")
-  public ResponseEntity<String> actualizarPersonaje(@RequestBody Personaje personaje, @PathVariable Integer id) {
+  public ResponseEntity<String> actualizarPersonaje(@RequestBody Personaje personaje, @PathVariable Long id) {
     personaje.setPeliculasSeries(getPeliculasDelRepositorio(personaje));
     personajeService.actualizar(personaje, id);
     return ResponseEntity.ok().body("El personaje fue actualizado correctamente");
@@ -72,7 +72,7 @@ public class PersonajeController {
 
   @DeleteMapping("/characters/{id}")
   @ApiOperation("Permite eliminar un personaje")
-  public ResponseEntity<String> eliminarPersonaje(@PathVariable Integer id) {
+  public ResponseEntity<String> eliminarPersonaje(@PathVariable Long id) {
     personajeService.eliminar(id);
     return ResponseEntity.ok().body("El personaje fue eliminado correctamente");
   }
