+++ src/main/java/com/example/challengebackendjava/model/Entidad.java
@@ -4,23 +4,23 @@ import com.example.challengebackendjava.serializer.View;
 import com.fasterxml.jackson.annotation.JsonView;
 
 abstract public class Entidad {
-    Integer id;
+  Long id;
 
-    @JsonView({
-            View.Personaje.Lista.class,
-            View.PeliculaSerie.Lista.class,
-            View.PeliculaSerie.Detalle.class,
-            View.Personaje.Detalle.class
-    })
-    public Integer getId() {
-        return id;
-    }
+  @JsonView({
+      View.Personaje.Lista.class,
+      View.PeliculaSerie.Lista.class,
+      View.PeliculaSerie.Detalle.class,
+      View.Personaje.Detalle.class
+  })
+  public Long getId() {
+    return id;
+  }
 
-    public void setId(Integer id) {
-        this.id = id;
-    }
+  public void setId(Long id) {
+    this.id = id;
+  }
 
-    abstract public void update(Entidad entidad);
+  abstract public void update(Entidad entidad);
 
-    abstract public boolean esValido();
+  abstract public boolean esValido();
 }
